separating core functionality
